siteblog/requirements.txt
asgiref==3.4.1
Django==3.2.6
django-autoslug==1.9.8
django-ckeditor==6.1.0
django-debug-toolbar==3.2.2
django-js-asset==1.2.2
django-rest-framework==0.1.0
djangorestframework==3.13.1
Pillow==9.0.1
pytz==2021.1
sqlparse==0.4.1
